requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
scikit-learn==1.2.2
streamlit==1.37.0